c09-Profiling Applications for Performance/requirements.txt
memory_profiler
pytest-benchmark